requirements.txt
jinja2
paho-mqtt
prometheus_client
requests
